Fix hono issue
