+++ front/pages/api/w/[wId]/credits/members-usage.ts
@@ -3,17 +3,61 @@
 /** @ignoreswagger */
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { GetMembersUsageResponseBody } from "@app/lib/api/credits/members_usage";
-import { handleGetMembersUsageRequest } from "@app/lib/api/credits/members_usage";
+import {
+  getMembersUsage,
+  MembersUsagePaginationSchema,
+} from "@app/lib/api/credits/members_usage";
 import type { Authenticator } from "@app/lib/auth";
+import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
 import type { NextApiRequest, NextApiResponse } from "next";
+import { fromError } from "zod-validation-error";
 
 async function handler(
   req: NextApiRequest,
   res: NextApiResponse<WithAPIErrorResponse<GetMembersUsageResponseBody>>,
   auth: Authenticator
 ): Promise<void> {
-  return handleGetMembersUsageRequest(req, res, auth);
+  if (!auth.isAdmin()) {
+    return apiError(req, res, {
+      status_code: 403,
+      api_error: {
+        type: "workspace_auth_error",
+        message: "Only workspace admins can access the members usage list.",
+      },
+    });
+  }
+
+  switch (req.method) {
+    case "GET": {
+      const paginationRes = MembersUsagePaginationSchema.safeParse(req.query);
+      if (!paginationRes.success) {
+        return apiError(req, res, {
+          status_code: 400,
+          api_error: {
+            type: "invalid_request_error",
+            message: `Invalid pagination parameters: ${fromError(paginationRes.error).toString()}`,
+          },
+        });
+      }
+
+      const body = await getMembersUsage({
+        auth,
+        paginationParams: paginationRes.data,
+        currentUrl: req.url ?? "/",
+      });
+
+      return res.status(200).json(body);
+    }
+    default:
+      return apiError(req, res, {
+        status_code: 405,
+        api_error: {
+          type: "method_not_supported_error",
+          message: "The method passed is not supported, GET is expected.",
+        },
+      });
+  }
 }
 
 export default withSessionAuthenticationForWorkspace(handler);
